增加了一些功能
一键大个子/小不点
一键一家鼠/土龙节节形态EC
增加了一些朱紫精灵一键生成多形态

+++ WangPluginPkm/WangUtil/NoexesCore/Noexes/Util/NoexsCommand.cs
@@ -11,7 +11,7 @@ namespace Noexes.Base
         {
             var cmd = GetCommands(NoexsCommandsEnum.Poke8);
             cmd.AddRange(BitConverter.GetBytes(address));
-            cmd.AddRange(BitConverter.GetBytes(value));
+            cmd.AddRange(BitConverter.GetBytes((short)value));
             return cmd.ToArray();
         }
 
@@ -19,7 +19,7 @@ namespace Noexes.Base
         {
             var cmd = GetCommands(NoexsCommandsEnum.Poke16);
             cmd.AddRange(BitConverter.GetBytes(address));
-            cmd.AddRange(BitConverter.GetBytes(value));
+            cmd.AddRange(BitConverter.GetBytes((short)value));
             return cmd.ToArray();
         }
